Handling nested sub state for Op-panel
This commit implement changes to handle the case of nested sub states.
Specifically, functionality 02(Select IPL type, system operating mode,
and firmware mode) and functionality 55(View and change platform system
dump data) needs a combination of values.

Here every next value is in continuation with the previous value selected
and a combination of all those values are required to execute the
functionality.

Every value in the set of values mentioned above is being referred as a
sub state and collectively as nested sub states.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>