Implementation of functionality 01
This commit reads system parameter like OSIPLType, systemOperatingMode,
HMCManaged, fwIPLType, hypervisor type over Dbus.

Some of the use case of reading these values are:
a) These values are required to create the output to be displayed on
LCD display with respect to execution of functionality 01 of Op-panel.

b) These values are also required to set the initial state of Op-panel
state manager when functionality 02 is entered.

Sample output when not HMC managed:
L1 : 01  D  N    PVM
L2 :             T

Change-Id: I5df2c76016cdb294ffb0f12e9b1821096168871d
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>