Transport Class Implementation (#14)
* Segmentation fault fix in panel

The button handler class closes the file descriptor at the destructor.

When io->run is passed after try catch, the file descriptor attached
to the button handler's stream decriptor is already closed, which causes
segmentation fault at the time of asynchronous read call.

Fix:
Moving io->run() into try catch block so that the button handler object
will never go out of scope and the file descriptor will remain opened.

Change-Id: I99270c184dc087f34fc9b66cf894de023b3b374e
Signed-off-by: PriyangaRamasamy <priyanga24@in.ibm.com>

* Panel:Transport class implementation

The BMC sends the commands to the panel micro controller
using raw i2c writes. The transport class handles the
communication between BMC and the micro controller.

This commit also implements logic to get right dbus path,
based on the system IM and device path (base/lcd)
and activate the transport class based on the panel presence.

When the panel is unmounted before firing the match signal,
and the transport key still remains as true, then the key validness
is lost. There occurs race conditon.

The commit checks for the panel presence before and after firing
the match signal and updates the key based on it.

Test:
Tested on simics and verified that when the present property is
getting changed the property change signal is captured by the panel
app and the transport key corresponding to the panel is activated to use
the transport class.

Change-Id: Ifa05b4f39e185d4f106464ab02eccd97a382c206
Signed-off-by: PriyangaRamasamy <priyanga24@in.ibm.com>