Function 13-SRC display (hex words 6 - 9)
This commit implements function 13.

This function displays a system reference code (SRC) on the
control panel to serve as a diagnostic aid that helps you
determine the source of a hardware or operating system problem.

Tested with dummy input for event id:
"BD8D1001 00000055 2E2D0010 00000000 00000000 00000000 00000000
00000000 00000000"

Sample output:
L1 : 0000000000000000
L2 : 0000000000000000

Change-Id: I4ebb21c3ff8117ea13d50bd59f2889422e54a8d2
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>