Implementation of Function 64
Sub functions associated with function 64 is used to display
data of the last 25 diagnostic status SRCs.

Enter a sub-function, in hexadecimal, between 00 and 18 to
look at the SRCs in sequential order.

In case the number of SRCs are less than 25, the count of sub
functions is de-activated accordingly.

If no SRCs exist, then only sub-function 00 will be available
and it will not display any SRC when selected.

Change-Id: Ic241263927383f5a2627e0ceaeba64ca7c664b5c
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>