Display 00/FF
Display FF in lcd panel to show that the triggerred function
failed. Function failure cases are specific to the functions.

If the function triggers a dbus call and it fails, then
that function execution is considered as a failure and we show
FF on panel.
If the function execution succeeds, we display 00 to indicate
success.

Signed-off-by: Priyanga Ramasamy <priyanga24@in.ibm.com>

Change-Id: Iff47851042825e01040442283a73ec87a0730c40