Implementation of function 25 and 26
This commit implements function 25 and 26 which are referred as
"Service Switch 1" and "Service Switch 2" respectively.

The Service Switch functions are only available in manual
operating mode when activated. Function 25 (Service Switch 1) and
Function 26 (Service Switch 2) can be used to enable or disable
the CE functions scroll range.

When the CE mode is disabled, a Service Switch 1 followed by a
Service Switch 2 enables that.

When the CE mode is enabled, execution of either function 25 or
function 26 disable the mode.

Change-Id: I5e5e0534bdade385f7395ad48fee34c7b5b8af78
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>