Panel to PHYP communication via PLDM
When there occurs a button press for a function
that needs to be executed by PHYP, the panel code
immediately sends the function number to PHYP that
needs to be executed. The communication between the
panel(from BMC) and the PHYP takes place using PLDM
protocol.

Test:
Tested on rainier.
Able to send the request message packet to PHYP
via PLDM.

Change-Id: I3c0a938e55192c4c32deb7b6935351304ad628d1
Signed-off-by: Priyanga Ramasamy <priyanga24@in.ibm.com>