Fetch existing PELs
This commit implement changes to fetch list of any existing
PEL that is present in the system when panel process comes up.

If found it sorts the list to filter last 25 PELs logged in
the system of desired severity, enables panel function w.r.t.
that and stores event ID of those PELs.

Change-Id: If932ffad3c4586f9f7bac9be52e5b4c6fc9010c5
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>