Panel function 74 implementation (#75)
* Panel to PHYP communication via PLDM (#58)

When there occurs a button press for a function
that needs to be executed by PHYP, the panel code
immediately sends the function number to PHYP that
needs to be executed. The communication between the
panel(from BMC) and the PHYP takes place using PLDM
protocol.

Test:
Tested on rainier.
Able to send the request message packet to PHYP
via PLDM.

Change-Id: I3c0a938e55192c4c32deb7b6935351304ad628d1
Signed-off-by: Priyanga Ramasamy <priyanga24@in.ibm.com>

Co-authored-by: Priyanga Ramasamy <priyanga24@in.ibm.com>

* Fixing clang errors in panel_app_test (#62)

Co-authored-by: giridhari <giridhari@giridharis-MacBook-Pro.local>

* Code fix for Function02 and Function01 (#60)

The commit adresses following points:
a) Pending TODO to read next boot side parameter for display
in function 01.

b) Fix to read correct parameter for boot type.

c) Fix for dbus write call in case of function 02.

d) Pending TODO to show FF in case of function 02 failure.

Change-Id: Ic04ccb27f5559dc9e94726f8b683e6efc2824efb
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>

* Generic GetPDR method (#65)

This commit replaces getPanelStateEffecterPDR method
with a common getPDR method which can return both the
pldm state effecter pdrs and sensor pdrs for any type
of pldm entity.

Change-Id: I10f30e679e0ce81604c27145a203cdbd94522601
Signed-off-by: Priyanga Ramasamy <priyanga24@in.ibm.com>

* Executor method to trigger PHYP functions (#68)

This commit implements a method in executor class
to trigger those functions executed by PHYP, whenever
there occurs a button press for such functions.

Change-Id: Icacac6ea763584f523b89e6b290880c1d503da74
Signed-off-by: Priyanga Ramasamy <priyanga24@in.ibm.com>

Co-authored-by: Priyanga Ramasamy <priyanga24@in.ibm.com>

* PEL terminating bit handle (#69)

This commit implements change to handle terminating bit in
the SRC of PELs logged by BMC.
In case the bit is set the SRC is directly sent to the LCD
display.

Change-Id: I740e7d254e1a6f85dbb88c28f831706956d48d83
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>

* Implementation of function 25 and 26 (#71)

This commit implements function 25 and 26 which are referred as
"Service Switch 1" and "Service Switch 2" respectively.

The Service Switch functions are only available in manual
operating mode when activated. Function 25 (Service Switch 1) and
Function 26 (Service Switch 2) can be used to enable or disable
the CE functions scroll range.

When the CE mode is disabled, a Service Switch 1 followed by a
Service Switch 2 enables that.

When the CE mode is enabled, execution of either function 25 or
function 26 disable the mode.

Change-Id: I5e5e0534bdade385f7395ad48fee34c7b5b8af78
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>

* Fix for Lamp test issue (#73)

Issue:
Missing 2 bytes of data in the lamp test command.
(lamp test period and lamp test on period).

This commit has the updated lamp test command.

Change-Id: Ie8fbf57f248178cb377402e9e4973f4798e50313
Signed-off-by: Priyanga Ramasamy <priyanga24@in.ibm.com>

* Panel function 74 implementation

The function on execution triggers a timer for a given
duration and sets Dbus property to mark window to upload
ACF certificate as active.
This property should be looked upon by Bmcweb to detect
if the window has been enabled.

Upon timer expiration the window is set as disabled.

Change-Id: I51a890b48c38fc52d94ff336a469f89c95f26160
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>

Co-authored-by: Priyanga Ramasamy <priyanga24@in.ibm.com>
Co-authored-by: Giridhari Krishna <giridharikrishnan@gmail.com>
Co-authored-by: giridhari <giridhari@giridharis-MacBook-Pro.local>
Co-authored-by: Santosh Puranik <santosh.puranik@in.ibm.com>