Fix mask check in state manager
The state manager was not properly enabling/disabling functions
based on the system state. Instead of checking if the bits
in the mask corresponding to the function were set in the
system state, it was checking for a complete equality in
those two values.

This commit fixes it by properly checking for the bits
in the system state against the mask.

Tested:
Tested by toggling system mode (manual/normal) and verifying
that the right set of functions get enabled/disabled.

Signed-off-by: Santosh Puranik <santosh.puranik@in.ibm.com>