monitor: Fix PEL Handling
This commit fixes a few issues found in the testing of handling
PEL D-Bus objects.

* The PEL daemon can create entries such as .../logging/blockx for
blocking PELs. We need to ignore such entries from our
GetManagedObjects call as they are not real PEL entries.
Fixed this by adding a filter for objects starting with the path
`/xyz/openbmc_project/logging/entry/`
* The code uses stoi to convert the PEL ID in the leaf of the
D-Bus object path to an integer for sorting purposes. However,
stoi can throw if the leaf name cannot be interpreted as an integer.
Handled this case by using a local try...catch.

Tested:
Tested on a system with a blocking PEL entry and ensured that
the panel app no longer crashes.

Signed-off-by: Santosh Puranik <santosh.puranik@in.ibm.com>