Flow update to set current operating mode
This commit implement changes to update panel internal state
for system operating mode right at the time of setting it
from function 02, instead of waiting for the DBus signal to
update the same.

This change was done to avoid any delay which can be there
between setting the mode, signal receiving and processing the
signal callback.

Also, code to set property "QuiesceOnHwError" is removed as
operating mode setting no longer depends on it.

Change-Id: I93025d0bf56ccb4df9faf5bb7b9ae02676ad6b4f
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>