Handle PEL deletion and update function 64 flow
The commit implement changes to handle PEL deletion event,
where in case last stored PEL was deleted then function 11
to 19 will be set to disabled mode.

It also make changes to the flow of function 64, where only
after execution of function 64 the code will fetch and filter
the list of diagnostic SRCs required to be displayed via this
function.

Change-Id: I88a55a2fd5f8220672eef0b9477fa688c48a4aa3
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>