Fix enabling of function 14 to 19
As the count of call out was not adjusted w.r.t 0th index being
the first entry, there was an extra function getting enabled
in the call out SRC function list.

The commit fixes that issue.

Test:
- Restart the panel service.
- Refer to the latest PEL resolution data from GUI.
- Check on LDC panel, number of functons enabled between 14 to 19
should be equal to the count of resolution string for the latest
PEL with desired severity.

Change-Id: Ic66cfcbd5fcf62cef87a7c642cea01e3e4736ec5
Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>