Also check for checkstop terminating SRC
In the same SRC error status flags byte as the termination due to FW
flag that the code is already checking there is a hardware checkstop
flag. Also check that when looking for terminating SRCs so that checkstop
PELs also make it to the panel.

Tested:
Created a checkstop PEL:
    "Hex Word 5":               "80000000",

Saw it on the panel (uncommented the cout lines):
    ibm-panel[4127]: L1 : BD70E510
    ibm-panel[4127]: L2 :

Checked the other bit still works:
    "Hex Word 5":               "20000000",

    ibm-panel[4127]: L1 : 11002602
    ibm-panel[4127]: L2 :

Signed-off-by: Matt Spinler <spinler@us.ibm.com>