Disable of OS IPL mode in function-02
The commit omits display and editing of OS IPL mode via function-02
unless the property osIplMode is set as true in panel interface
via PLDM.

The property is set when the system has single partition with HMC
connected.

Test:
- Boot a system with the property osIplMode under panel interface
set as false.
- Launch function-02 on operator panel.
- Initial screen will not display OS IPL mode.

Implies in this mode user will not be allowed to edit OS IPL mode
via panel.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>