Fix system dump issue from OP-Panel
When the user initiates ‘system dump’ from OP-Panel on the system using
function 42, is not being triggered.

Reason for this is “panel” service is not sending required parameters
to ‘Dump manager’ service to initiate ‘system dump’, DBus call is
failed for the request made.

This commit adds the code to fix the above issue in the panel.

Signed-off-by: Anupama B R <anupama.b.r1@ibm.com>