Lcd panel: Fix "device not found" error #729129
The Panel app fails with the error "device not found" after an AC cycle
when it starts before the vpd-manager.

The vpd-manager is responsible for enabling the LCD microcontroller
I2C GPIO pin. When the Panel app starts after vpd-manager, it can
directly access the LCD microcontroller without issues.

However, if the Panel app comes up earlier, the GPIO pin is not yet
enabled, and access to the LCD microcontroller fails.

This change fixes the issue by checking the LCD microcontroller GPIO
presence pin and enabling the I2C GPIO pin before accessing the
microcontroller.

Signed-off-by: Anupama B R <anupama.b.r1@ibm.com>